Fix: 2 calls to AccessToken#serialize lose actor
When we have an actor, `encode_for_jwt` calls `encode_for_jwt` to
serialize the actor... but that method alters its input. As a result, we
end up modifying the actor hash that was passed in!

Since the hash is mutable, that modifies the same instance as `@actor`.
As a result, repeated calls to `#encode_for_jwt` (and therefore,
repeated calls to `#serialize`) will cause the `AccessToken` instance to
lose its actor.

We haven't historically hit this since our APIs don't typically
serialize the same token instance twice... but I hit this in specs,
where we do.

This fixes that.