Restrict cookies to only the auth domain
The Fridge cookie is used to store a read-only ("sessionized") token. It
can then be used to GET the `current_token` endpoint on auth to get a
"manage" scoped token.

We are restricting the token to the auth subdomain only so that our
users' browsers will not send it to ALL aptible.com subdomains, as it
currently does. We are now hosting some aptible subdomains with
third-pary vendors as well as in customer stacks for Crazylegs. This
change will prevent those services from getting access to Aptible tokens
through those cookies.