Allow for domains with ports specified if used on localhost
Right now if auth-api is run off of localhost, then it will try to set
that specific domain, which will fail because most browsers will ignore
it.

Previously we left it as the default domain, which browsers allow for
localhost cookies. I don't understand WHY, but that's my best googling
of the problem that Maryam encountered.

So now we strip off the port number to allow it be specified as well as
special casing handling of localhost. This should let local development
sessions work "correctly" but the behavior will be slightly different
than in production where the cookie set will NOT work for our multiple
services.

If folks need that behavior, they should ensure the the various APIs are
on different subdomains (e.g,. via /etc/hosts, docker-compose
networking, etc.)